+++ packages/terra-form-validation/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.89.0 - (October 11, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 1.88.0 - (October 3, 2023)
 
 * Changed